Added Husky and nano-staged. (#5)
I don't use biomejs, so I didn't know if I was formatting correctly or
linting. I had minimal defense with recomandation, but I thought it was
a negligible barrier (me? I'm not going to install it? lol).

So I created a barrier to make commits safe with husky and nano-staged.

- lint
- formatting
- type checking

---------

Co-authored-by: Jonghyeon Ko <manudeli.ko@gmail.com>