ci(turbo): update turbo version (#71)
# Overview

I think the version of the "turbo" dependency specified as "latest" in
the "package.json" is meaningless. The version is locked to the latest
version (1.13.3) at the time this project was created, as defined in the
"pnpm-lock.yaml" file. In this project, "latest" does not point to the
latest version, but rather to the latest version at the time the lock
file was created.

It would be less confusing for contributors if the version to be used
was explicitly specified. While it is convenient to update the version
using "pnpm update turbo", one has to check the "node_modules" or
"pnpm-lock.yaml" file to determine the exact version.

If we had specified version 1.13.3 or ^1.13.3, we could have easily
understood that it referred to versions less than 2.0.0. However,
currently, we have no idea if it's version 1 or version 2. Moreover, the
latest version of "turbo" is 2.0.5.

Can you help reduce the effort developers need to find the version?
Below is the terminal UI officially provided by turbo version 2. It has
great visual effects and allows users to check the progress through
keyboard interaction.


https://github.com/coauthors/coauthors/assets/55759551/bc8014b5-96bc-4430-8455-ec96f43221d4


## PR Checklist

- [X] I did below actions if need

1. I read the [Contributing
Guide](https://github.com/coauthors/coauthors/blob/main/CONTRIBUTING.md)
2. I added documents and tests.